# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Crewquarters Control API
  description: 'Control plane for Crewquarters. JSON is camelCase. Errors use `{error: {code, message,
    requestId, details}}`. State-changing requests need the session cookie, an allowed Origin, and the
    `X-CSRF-Token` header. Mutating requests accept an `Idempotency-Key` header.'
  version: 0.1.0
paths:
  /api/v1/bootstrap:
    post:
      tags:
      - auth
      summary: Create the device owner with the one-time bootstrap token
      operationId: bootstrap_api_v1_bootstrap_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/agent-installations:
    get:
      tags:
      - agents
      summary: List installed agents
      operationId: list_installations_api_v1_agent_installations_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_InstallationOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - agents
      summary: Install an agent, approving its exact permissions
      operationId: create_installation_api_v1_agent_installations_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationCreateIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
  /api/v1/agent-installations/{installation_id}:
    get:
      tags:
      - agents
      summary: Get an installed agent with readiness
      operationId: get_installation_api_v1_agent_installations__installation_id__get
      parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Installation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    patch:
      tags:
      - agents
      summary: Update configuration, enablement, version, or permission approval
      operationId: update_installation_api_v1_agent_installations__installation_id__patch
      parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Installation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationPatchIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    delete:
      tags:
      - agents
      summary: Uninstall an agent (soft delete; blocked while runs are active)
      operationId: delete_installation_api_v1_agent_installations__installation_id__delete
      parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Installation Id
      responses:
        '204':
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
...
  /api/v1/runs/{run_id}:
    get:
      tags:
      - runs
      summary: Get a run
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
...
components:
  schemas:
...
    BootstrapIn:
      properties:
        token:
          type: string
          maxLength: 256
          minLength: 16
          title: Token
          description: One-time bootstrap token.
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: ^[A-Za-z0-9_.@-]+$
          title: Username
        password:
          type: string
          maxLength: 1024
          minLength: 12
          title: Password
        email:
          anyOf:
          - type: string
            maxLength: 254
          - type: 'null'
          title: Email
      type: object
      required:
      - token
      - username
      - password
      title: BootstrapIn
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          examples:
          - MODEL_CAPACITY_EXCEEDED
        message:
          type: string
          title: Message
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
        details:
          additionalProperties: true
          type: object
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InstallationCreateIn:
      properties:
        agentId:
          type: string
          title: Agentid
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
          description: Defaults to the catalog's current version.
        config:
          additionalProperties: true
          type: object
          title: Config
        approvedPermissions:
          additionalProperties: true
          type: object
          title: Approvedpermissions
          description: Must equal the version's requested permissions exactly.
        modelBindings:
          additionalProperties:
            type: string
          type: object
          title: Modelbindings
          description: Owner choice of variant per requested profile family.
          examples:
          - local.general: local.general.quality
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - agentId
      - approvedPermissions
      title: InstallationCreateIn
    InstallationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        agentId:
          type: string
          title: Agentid
        agentName:
          type: string
          title: Agentname
        agentVersion:
          type: string
          title: Agentversion
        agentVersionId:
          type: string
          format: uuid
          title: Agentversionid
        config:
          additionalProperties: true
          type: object
          title: Config
        requestedPermissions:
          additionalProperties: true
          type: object
          title: Requestedpermissions
        approvedPermissions:
          additionalProperties: true
          type: object
          title: Approvedpermissions
        capabilities:
          items:
            type: string
          type: array
          title: Capabilities
        modelBindings:
          additionalProperties:
            type: string
          type: object
          title: Modelbindings
        needsReapproval:
          type: boolean
          title: Needsreapproval
        enabled:
          type: boolean
          title: Enabled
        version:
          type: integer
          title: Version
        readiness:
          $ref: '#/components/schemas/Readiness'
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - agentId
      - agentName
      - agentVersion
      - agentVersionId
      - config
      - requestedPermissions
      - approvedPermissions
      - capabilities
      - modelBindings
      - needsReapproval
      - enabled
      - version
      - readiness
      - createdAt
      - updatedAt
      title: InstallationOut
    InstallationPatchIn:
      properties:
        version:
          type: integer
          title: Version
          description: Current installation version (optimistic concurrency).
        agentVersion:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentversion
          description: Upgrade/downgrade to this agent version.
        config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config
        approvedPermissions:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Approvedpermissions
        modelBindings:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Modelbindings
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      type: object
      required:
      - version
      title: InstallationPatchIn
...
    Page_InstallationOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/InstallationOut'
          type: array
          title: Items
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
          description: Opaque cursor for the next page; null on the last page.
      type: object
      required:
      - items
      title: Page[InstallationOut]
...
    Readiness:
      properties:
        ready:
          type: boolean
          title: Ready
        checks:
          items:
            $ref: '#/components/schemas/ReadinessCheck'
          type: array
          title: Checks
      type: object
      required:
      - ready
      - checks
      title: Readiness
    ReadinessCheck:
      properties:
        name:
          type: string
          enum:
          - enabled
          - permissions
          - configuration
          - architecture
          - connection
          - model
          title: Name
        status:
          type: string
          enum:
          - ok
          - missing
          - needs_attention
          title: Status
        detail:
          type: string
          title: Detail
        resource:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource
      type: object
      required:
      - name
      - status
      - detail
      title: ReadinessCheck
...
    RunOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        installationId:
          type: string
          format: uuid
          title: Installationid
        agentId:
          type: string
          title: Agentid
        agentName:
          type: string
          title: Agentname
        agentVersion:
          type: string
          title: Agentversion
        trigger:
          type: string
          enum:
          - manual
          - schedule
          title: Trigger
        scheduleId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scheduleid
        scheduledFor:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduledfor
        state:
          type: string
          enum:
          - QUEUED
          - PREPARING
          - LOADING_MODEL
          - RUNNING
          - WAITING_INPUT
          - CANCELLING
          - SUCCEEDED
          - FAILED
          - CANCELLED
          - INTERRUPTED
          title: State
        currentAttempt:
          type: integer
          title: Currentattempt
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        error:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Error
        retryable:
          type: boolean
          title: Retryable
        cancelRequested:
          type: boolean
          title: Cancelrequested
        acknowledgedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acknowledgedat
        activeSecondsUsed:
          type: number
          title: Activesecondsused
        inputWaitSecondsUsed:
          type: number
          title: Inputwaitsecondsused
        activeTimeoutSeconds:
          type: integer
          title: Activetimeoutseconds
        maxInputWaitSeconds:
          type: integer
          title: Maxinputwaitseconds
        usesCloud:
          type: boolean
          title: Usescloud
        pendingInputCount:
          type: integer
          title: Pendinginputcount
        createdAt:
          type: string
          format: date-time
          title: Createdat
        startedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startedat
        finishedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finishedat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - installationId
      - agentId
      - agentName
      - agentVersion
      - trigger
      - scheduleId
      - scheduledFor
      - state
      - currentAttempt
      - result
      - error
      - retryable
      - cancelRequested
      - acknowledgedAt
      - activeSecondsUsed
      - inputWaitSecondsUsed
      - activeTimeoutSeconds
      - maxInputWaitSeconds
      - usesCloud
      - pendingInputCount
      - createdAt
      - startedAt
      - finishedAt
      - updatedAt
      title: RunOut
...
    SessionOut:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        csrfToken:
          type: string
          title: Csrftoken
          description: Send as the X-CSRF-Token header on state-changing requests.
        expiresAt:
          type: string
          format: date-time
          title: Expiresat
        idleExpiresAt:
          type: string
          format: date-time
          title: Idleexpiresat
      type: object
      required:
      - user
      - csrfToken
      - expiresAt
      - idleExpiresAt
      title: SessionOut
...
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        role:
          type: string
          enum:
          - owner
          - member
          title: Role
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - username
      - email
      - role
      - createdAt
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: cq_session
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
    service:
      type: http
      scheme: bearer
      description: Internal service credential
servers:
- url: /
  description: Same origin as the web UI